+++ packages/discourse_ui/lib/views/private_messaging/conversation/widgets/conversation_header_widget.dart
@@ -165,7 +165,7 @@ class ConversationHeaderWidget extends StatelessWidget {
                               : BlendMode.color,
                           ),
                           child: Image.asset(
-                            'assets/forum_header_bg.png',
+                            'packages/discourse_ui/assets/forum_header_bg.png',
                             fit: BoxFit.cover,
                             width: double.infinity,
                             height: double.infinity,